fix demo pages not being scrollable
